+++ Reqnroll/EnvironmentAccess/IBuildMetadataProvider.cs
@@ -1,0 +1,7 @@
+﻿namespace Reqnroll.EnvironmentAccess
+{
+    public interface IBuildMetadataProvider
+    {
+        BuildMetadata GetBuildMetadata();
+    }
+}